Move code from genesis.rs to chain_spec.rs
