+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
 	"nodes/meta-node",
 	"example/example",
 ]
-librocksdb-sys = "6.17.3"
 
 [patch.crates-io]
 syn = { git = 'https://github.com/dtolnay/syn', tag = '1.0.97', version = '1.0.97' }
