+++ runtimes/meta-runtime/src/lib.rs
@@ -8,10 +8,6 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
-// Include the genesis helper module when building to std
-#[cfg(feature = "std")]
-pub mod genesis;
-
 use pallet_transaction_payment::CurrencyAdapter;
 use sp_api::impl_runtime_apis;
 use sp_core::{OpaqueMetadata, H256};
